requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
tortoise-orm==0.19.3
asyncpg==0.28.0